requirements.txt
deepgram-sdk>=3.5.0
websockets>=12.0
openai>=1.12.0
anthropic>=0.20.0
python-dotenv>=1.0.1
